+++ src/js/components/internal/slider-preload.js
@@ -2,18 +2,17 @@ import { $$, observeIntersection, removeAttr } from 'uikit-util';
 
 export default {
     connected() {
-        this.observer = observeIntersection(
-            this.slides.concat(this.$el),
-            () => {
-                for (const el of this.getAdjacentSlides()) {
-                    el && $$('img[loading="lazy"]', el).forEach((el) => removeAttr(el, 'loading'));
-                }
-            },
-            { rootMargin: '50% 50%' }
+        this.registerObserver(
+            observeIntersection(
+                this.slides.concat(this.$el),
+                (entries, observer) => {
+                    for (const el of this.getAdjacentSlides()) {
+                        $$('img[loading="lazy"]', el).forEach((el) => removeAttr(el, 'loading'));
+                        observer.unobserve(el);
+                    }
+                },
+                { rootMargin: '50% 50%' }
+            )
         );
-    },
-
-    disconnected() {
-        this.observer.disconnect();
     },
 };